// src/app/bna-trainee-biodata/biodata-tab-layout/models/EducationalInstitution.ts
export interface EducationalInstitution {
    educationalInstitutionId: number;
    traineeId: number;  
    instituteName: string;    
    address: string;
    districtId: number;  
    district: string; 
    thanaId: number;
    thana: string; 
    classStudiedFrom: string;  
    classStudiedTo: string;
    yearFrom: string;
    yearTo: string;
    additionaInformation: string;    
    status: number;
    menuPosition: number;
    isActive: boolean;
} 